Extend StatusEntry.supported_media_commands from u8 to u32. (#6)
ChromeCast devices started to return supportedMediaCommand with high
bits set which won't fit into u8. Trying to parse media status messages
would therefor return Err(..).

+++ src/cast/proxies.rs
@@ -264,7 +264,7 @@ pub mod media {
         #[serde(rename = "currentTime")]
         pub current_time: Option<f32>,
         #[serde(rename = "supportedMediaCommands")]
-        pub supported_media_commands: u8,
+        pub supported_media_commands: u32,
     }
 
     #[derive(Deserialize, Debug)]
